requirements.txt
Django==1.10
django-rest-framework-mongoengine==3.3.0
djangorestframework==3.4.4
ecdsa==0.13
factory-boy==2.7.0
fake-factory==0.6.0
future==0.15.2
ipaddress==1.0.16
mongoengine==0.10.6
py==1.4.31
pycrypto==2.6.1
pymongo==3.3.0
pytest==2.9.2
pytest-django==2.9.1
python-dateutil==2.5.3
python-jose==1.2.0
six==1.10.0
wheel==0.24.0
